fixed: compile warning.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@31481 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/paplayer/PAPlayer.cpp
@@ -869,7 +869,7 @@ void PAPlayer::Seek(bool bPlus, bool bLargeStep)
 void PAPlayer::SeekTime(__int64 iTime /*=0*/)
 {
   if (!CanSeek()) return;
-  int seekOffset = iTime - GetTime();
+  int seekOffset = (int)(iTime - GetTime());
   if (m_currentFile->m_lStartOffset)
     iTime += m_currentFile->m_lStartOffset * 1000 / 75;
   m_SeekTime = iTime;
